Move NMS code to image processor
